+++ frontend/ogy_dashboard/src/pages/account/available-ogy/index.tsx
@@ -25,8 +25,8 @@ const AvailableOGY = () => {
       <div className="flex justify-between mb-4">
         <div className="font-bold text-content/60">Available OGY</div>
         <Link
-          className="font-semibold text-accent hover:underline"
-          to={`/explorer/transactions/accounts/${principal}`}
+          className="text-accent"
+          to={`/explorer/transactions/accounts/${principal}/history`}
         >
           Transaction history
         </Link>